修正一些BUG

+++ 2LowLayer/USART/bsp_uart_fifo.c
@@ -689,9 +689,9 @@ void RS485_ReciveNew(uint8_t _byte)
 //	MODBUS_ReciveNew(_byte);
 }
 
-uint8_t RS485_Recv(COM_PORT_E _ucPort,uint8_t *buf, uint8_t len)
+uint16_t RS485_Recv(COM_PORT_E _ucPort,uint8_t *buf, uint16_t len)
 {
-    uint8_t i = 0;
+    uint16_t i = 0;
 
     UART_T *pUart;
 	pUart = ComToUart(_ucPort);
@@ -715,11 +715,11 @@ uint8_t RS485_Recv(COM_PORT_E _ucPort,uint8_t *buf, uint8_t len)
 }
 
 
-uint8_t RS485_RecvAtTime(COM_PORT_E _ucPort,uint8_t *buf, uint8_t len,uint32_t timeout)
+uint16_t RS485_RecvAtTime(COM_PORT_E _ucPort,uint8_t *buf, uint16_t len,uint32_t timeout)
 {
     //uint8_t i = 0;   
-    uint8_t recvSize = len;
-    uint8_t recvLen = 0;
+    uint16_t recvSize = len;
+    uint16_t recvLen = 0;
     //uint8_t tmp[1] = {0};
     UART_T *pUart;
 	pUart = ComToUart(_ucPort);
